Fix: obrigar o preenchimento obrigatório de matemática do 4 ao a 9 ano
